Moved the BaseTestCase and SilentReporter
I moved them to the utils folder. Inspired by how we do things in
VoIPGRID.
